requirements.txt
fastapi
uvicorn[standard]
python-dotenv
paymanai-sdk
requests
